changes after CR
